pyproject.toml
[project]
name = "migration-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "crewai==1.9.3",
    "crewai-tools==1.9.3",
    "ipykernel>=7.2.0",
    "litellm>=1.75.3",
    "openai>=1.83.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
